+++ src/test/test_bloomfilter.c
@@ -67,7 +67,7 @@ CTEST2(bloom, create)
 	ASSERT_EQUAL(INTSIZE, data->b1->size);
 
 	int x = 0;
-	for (int i = 0; i < data->b1->size; ++i)
+	for (size_t i = 0; i < data->b1->size; ++i)
 	{
 		x |= data->b1->a[i];
 	}
